refactor: findAlluser

+++ back-end/src/services/admin/deleteUser.js
@@ -2,8 +2,10 @@ const { users } = require('../../database/models');
 const { Unauthorized, notFound } = require('../../utils/statusCode');
 const errorConstructor = require('../../utils/errorHandling');
 
-module.exports = async (adminRole, id) => {
+module.exports = async (adminRole, adminId, id) => {
   if (adminRole !== 'administrator') throw errorConstructor(Unauthorized, 'unauthorized user');
+  console.log(adminId, id);
+  if (Number(adminId) === Number(id)) throw errorConstructor(Unauthorized, 'deletion not allowed');
 
   const excluded = await users.destroy({ where: { id } });
   if (excluded === 0) throw errorConstructor(notFound, 'User not found');
